+++ pykit/objinfo/test_objectinfo.py
@@ -3,7 +3,7 @@ from pykit.objinfo._info import ObjectInfo
 
 def test_main():
     info: ObjectInfo = ObjectInfo[int](("hello", 2))
-    assert str(info) == "hello=<2>"
+    assert str(info) == "hello 2"
 
 
 def test_only_title():